+++ src/build_event/build_event_for_data.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List
 
 from ..fetch_data.polymarket.polymarket_client import PolymarketClient
 
-from ..utils.config_loader import Config
+from ..core.config import Config
 
 @dataclass
 class BuildedPmEvent:
