+++ src/core/filters/__init__.py
@@ -1,0 +1,1 @@
+# Filters module - signal filtering system
