+++ src/monitors/pnl_monitor.py
@@ -18,10 +18,10 @@ from typing import Optional
 
 from ..api.pnl import get_pnl_summary
 from ..telegram.TG_bot import TG_bot
-from ..utils.config_loader import load_all_configs
+from ..core.config import load_all_configs
 from ..utils.SqliteHandler import SqliteHandler
 from ..utils.state_tracker import check_state_completed, mark_state_completed, get_state_key
-from ..utils.save_data.save_pnl_snapshot import PnlSnapshot
+from ..core.save.save_pnl_snapshot import PnlSnapshot
 
 logger = logging.getLogger(__name__)
