+++ src/services/execute_trade.py
@@ -7,8 +7,8 @@ from ..telegram.TG_bot import TG_bot
 from ..trading.deribit_trade import DeribitUserCfg
 from ..trading.deribit_trade_client import Deribit_trade_client
 from ..trading.polymarket_trade_client import Polymarket_trade_client
-from ..utils.config_loader import Env_config
-from ..utils.save_data.save_position import save_position
+from ..core.config import Env_config
+from ..core.save.save_position import save_position
 
 logger = logging.getLogger(__name__)
